Added xRewritefileln, BUT works not really as advertised.
Tests have shown it to not work bi-atomically as expected: a reader using
 my $f; sysopen $f, "foo", O_RDONLY or die $!;
 my $cnt; defined (sysread( $f,$cnt,100)) or die $!;
 close $f or die $!;
with two parallel writers can see the output of one writer overlaid
over the other; of course, when the sequence is A(truncates),
B(truncates), B(writeslong), A(writesshort).

So far if one only considers the contents up to a predefined delimter
(newline?), then it seems to work. But then the truncate is
essentially of no use.