source .bash_profile for PATH (and possibly other important setup stuff)
Emacs should get the same environment as when starting it from a
shell, that's why I'm source'ing .bash_profile. Of course this now
assumes that the user is using bash. Oh well. What alternative would
there be? If it's still to be per-user configurable? Yep, check which
shell the user is using; basically, run the whole thing from a
shell. But, that should Gnome be doing really, shouldn't it, the whole
Gnome session should be started through such a shell instance, or not?