Remove, since exactly it did lower the limit now and made me crash
Well, and it bothered with-gdb-backtrace-to. (But the segfault
probably was because of the out of memory I guess. Well, to be honest
emacs said it ran out of mem several times -- upon opening binary
files; so maybe just this time it didn't go well 'cause after gc still
not enough memory or so.)

Only bad thing is I don't know what to do on servers where the default
ulimit is actually lower. (Bash function instead? Wouldn't work when
starting emacs indirectly like, e.g., from with-gdb-backtrace-to --
well..)