(change to be able to pass arguments)
until I realize that the loop running the command is in another child,
and signalled by a unix signal.

Yes, this was intended to be used to serialize run of a command. And
not for remote (as-another-user) execution as I want to do now. sigh?

It was intentionally 'lossy'. (Not using RT signals.)