Wrapper around sort to force LANG=C (CRACK)
Is this the first time I run into this issue? I don't think so but I'm
sure I had forgotten and now just spent another half hour to find out
just how to do it. Don't ask for how risky?

Now, should we rather give a warning if LANG is not C ?
Maybe?