Mylock: fix claim feature
This assumes that, claiming only ever happens when locks are stale
(not that the timeout is too short / processes are being suspended and
then resumed).

Also, relies on intermediate lock files in the chain of lock files not
vanishing (but OSes may not guarantee link creation order over
crashes!).

So, not perfect yet.